requirements.txt
aiosqlite==0.18.0
discord.py==2.1.0
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
requests==2.28.1
tensorflow==2.11.0
tensorflow_macos==2.11.0
